Updated JSON Structure
All tables are now represented in the json file with a "fields" JSONArray containing the values from the table. This means the fields are in the same order as in the database, making it much easier to upload plot packages without any manipulation.

This change required a lot of files being modified, but it's worth it in the long run. It makes it much easier to manage each table and its associated values.