Implemented JSON Mapping
Previously, the application was extracting plot packages from JSON files as a single JSON Object. This JSON Object was stored in the Package class and would be used to upload all tables to the database. This logic was in the Main Controller, but should be located within each Table class instead.

Now the JSON file data is correctly mapped to the Table classes, allowing the table upload logic to be performed within the class on its own fields.